backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis and Celery
redis[hiredis]==5.0.1
celery==5.3.4

# LangGraph and OpenAI
langgraph==0.0.60
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7

# Vector Database
pymilvus==2.3.4

# ML and NLP
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# Web scraping and content processing
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
scrapy==2.11.0
python-dateutil==2.8.2

# Text processing and validation
spacy==3.7.2
textstat==0.7.3
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Real-time communication
python-socketio==5.10.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.7